Allow custom providers for RecheckSeleniumAdapter
